requirements.txt
# Google Agent Development Kit
google-adk>=1.0.0

# LangChain core (base classes for custom LLMs)
langchain-core>=0.2.0

# HTTP client
requests>=2.31.0

# ── Mutual Fund Planner dependencies ──────────────────────────────────────

# AMFI official mutual fund data (NAV, scheme codes)
mftool>=2.2.0

# YouTube transcript extraction (CNBC, ET Now, etc.)
youtube-transcript-api>=0.6.0

# HTML parsing for Moneycontrol / ET / LiveMint scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Env file loading
python-dotenv>=1.0.0
